refactor: [#35] id 타입 변경 integer -> long

+++ src/main/java/com/campool/controller/GearTypeController.java
@@ -45,7 +45,7 @@ public class GearTypeController {
 
     @LoginValidation(role = Role.ADMIN)
     @DeleteMapping("/types/{id}")
-    public void deleteGearType(@PathVariable int id) {
+    public void deleteGearType(@PathVariable long id) {
         gearTypeService.deleteById(id);
     }
